requirements.txt
# Core dependencies
python-dotenv>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0

# RSS and content extraction
feedparser>=6.0.10
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2

# AI and NLP
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0

# Document generation
markdown2>=2.4.10
fpdf2>=2.7.5
pyyaml>=6.0.1

# Optional extensions
# sentence-transformers>=2.2.2
# faiss-cpu>=1.7.4
# apscheduler>=3.10.1
# streamlit>=1.25.0
# typer>=0.9.0